+++ shiki_organizer/gui/__main__.py
@@ -1,5 +1,6 @@
 import sys
 
+from PySide6.QtCore import Signal
 from PySide6.QtWidgets import QApplication, QMainWindow
 
 from shiki_organizer.gui.interval_tab import IntervalTab
@@ -10,12 +11,14 @@ from shiki_organizer.gui.ui.main_window import Ui_MainWindow
 
 
 class MainWindow(QMainWindow, Ui_MainWindow):
+    update_all = Signal()
+
     def __init__(self):
         super().__init__()
         self.setupUi(self)
-        self.tab_widget.addTab(TasksTab(True), "Today")
-        self.tab_widget.addTab(TasksTab(False), "Tasks")
-        self.tab_widget.addTab(IntervalTab(), "Intervals")
+        self.tab_widget.addTab(TasksTab(self.update_all, True), "Today")
+        self.tab_widget.addTab(TasksTab(self.update_all, False), "Tasks")
+        self.tab_widget.addTab(IntervalTab(self.update_all), "Intervals")
         self.tab_widget.addTab(RepositoryTab(), "Repository")
         self.tab_widget.addTab(SettingsTab(), "Settings")