+++ shiki_organizer/gui/task.py
@@ -13,10 +13,11 @@ class TaskModel(QAbstractTableModel):
     def __init__(self, today=True, parent=None):
         super().__init__(parent)
         self.today = today
-        tasks = Task.select().where(Task.archived == False)
+        self.tasks = Task.select().where(Task.archived == False)
         if self.today:
-            tasks = tasks.where(Task.scheduled <= dt.date.today())
-        self._tasks = [task for task in tasks]
+            self.tasks = self.tasks.where(Task.scheduled <= dt.date.today())
+        self.sort_column = 0
+        self.sort_order = Qt.SortOrder.DescendingOrder
         self._headers = [
             "ID",
             "Priority",
@@ -39,54 +40,85 @@ class TaskModel(QAbstractTableModel):
             if role == Qt.DisplayRole:
                 match index.column():
                     case 0:
-                        return self._tasks[index.row()].id
+                        return self.tasks[index.row()].id
                     case 1:
-                        return self._tasks[index.row()].priority
+                        return self.tasks[index.row()].priority
                     case 2:
-                        return self._tasks[index.row()].description
+                        return self.tasks[index.row()].description
                     case 3:
-                        if self._tasks[index.row()].recurrence:
-                            return str(self._tasks[index.row()].recurrence) + "d"
+                        if self.tasks[index.row()].recurrence:
+                            return str(self.tasks[index.row()].recurrence) + "d"
                     case 4:
-                        scheduled = self._tasks[index.row()].scheduled
+                        scheduled = self.tasks[index.row()].scheduled
                         return str(scheduled) if scheduled else ""
                     case 5:
-                        deadline = self._tasks[index.row()].deadline
+                        deadline = self.tasks[index.row()].deadline
                         return str(deadline) if deadline else ""
                     case 6:
-                        return self._tasks[index.row()].days
+                        return self.tasks[index.row()].days
                     case 7:
-                        return duration_to_str(self._tasks[index.row()].average)
+                        return duration_to_str(self.tasks[index.row()].average)
                     case 8:
-                        return duration_to_str(self._tasks[index.row()].duration)
-                    case _:
-                        return self._tasks[index.row()].id
+                        return duration_to_str(self.tasks[index.row()].duration)
             elif role == Qt.BackgroundRole:
-                if self._tasks[index.row()] == actions.get_current_task():
+                if self.tasks[index.row()] == actions.get_current_task():
                     return QBrush(QColor(0, 255, 0))
 
     def rowCount(self, parent: QModelIndex) -> int:
-        return len(self._tasks)
+        return len(self.tasks)
 
     def columnCount(self, parent: QModelIndex) -> int:
         return 9
 
     def refresh(self):
-        tasks = Task.select().where(Task.archived == False)
+        self.tasks = Task.select().where(Task.archived == False)
         if self.today:
-            tasks = tasks.where(Task.scheduled <= dt.date.today())
-        self._tasks = [task for task in tasks]
+            self.tasks = self.tasks.where(Task.scheduled <= dt.date.today())
+
+        def sort_by_any(asc, desc):
+            if self.sort_order == Qt.SortOrder.AscendingOrder:
+                return asc
+            else:
+                return desc
+
+        def sort_by_field(asc, desc):
+            return sort_by_any(self.tasks.order_by(asc), self.tasks.order_by(desc))
+
+        match self.sort_column:
+            case 0:
+                self.tasks = sort_by_field(Task.id, Task.id.desc())
+            case 1:
+                self.tasks = sort_by_field(Task.priority, Task.priority.desc())
+            case 2:
+                self.tasks = sort_by_field(Task.description, Task.description.desc())
+            case 3:
+                self.tasks = sort_by_field(Task.recurrence, Task.recurrence.desc())
+            case 4:
+                self.tasks = sort_by_field(Task.scheduled, Task.scheduled.desc())
+            case 5:
+                self.tasks = sort_by_field(Task.deadline, Task.deadline.desc())
+            case 6:
+                self.tasks = sort_by_field(Task.days, Task.days.desc())
+            case 7:
+                self.tasks = sort_by_any(
+                    sorted(self.tasks, key=lambda x: x.average),
+                    sorted(self.tasks, key=lambda x: x.average, reverse=True),
+                )
+            case 8:
+                self.tasks = sort_by_field(Task.duration, Task.duration.desc())
         self.layoutChanged.emit()
 
     def done(self, indexes):
         github_token = os.environ.get("GITHUB_TOKEN", "")
         actions.done([str(index.data()) for index in indexes], github_token)
-        self.refresh()
 
     def start(self, index):
         actions.start(str(index.data()), dt.datetime.now(), None)
-        self.refresh()
 
     def stop(self):
         actions.stop()
+
+    def sort(self, column: int, order: Qt.SortOrder) -> None:
+        self.sort_column = column
+        self.sort_order = order
         self.refresh()